+++ src/components/DayPickerSingleDateController.jsx
@@ -19,14 +19,16 @@ import toISODateString from '../utils/toISODateString';
 import toISOMonthString from '../utils/toISOMonthString';
 
 import ScrollableOrientationShape from '../shapes/ScrollableOrientationShape';
+import DayOfWeekShape from '../shapes/DayOfWeekShape';
 
 import {
   HORIZONTAL_ORIENTATION,
   VERTICAL_SCROLLABLE,
   DAY_SIZE,
-} from '../../constants';
+} from '../constants';
 
 import DayPicker from './DayPicker';
+import OutsideClickHandler from './OutsideClickHandler';
 
 const propTypes = forbidExtraProps({
   date: momentPropTypes.momentObj,
@@ -37,7 +39,6 @@ const propTypes = forbidExtraProps({
   onClose: PropTypes.func,
 
   keepOpenOnDateSelect: PropTypes.bool,
-  minimumNights: PropTypes.number,
   isOutsideRange: PropTypes.func,
   isDayBlocked: PropTypes.func,
   isDayHighlighted: PropTypes.func,
@@ -52,8 +53,12 @@ const propTypes = forbidExtraProps({
   orientation: ScrollableOrientationShape,
   withPortal: PropTypes.bool,
   initialVisibleMonth: PropTypes.func,
+  firstDayOfWeek: DayOfWeekShape,
   hideKeyboardShortcutsPanel: PropTypes.bool,
   daySize: nonNegativeInteger,
+  verticalHeight: nonNegativeInteger,
+  noBorder: PropTypes.bool,
+  transitionDuration: nonNegativeInteger,
 
   navPrev: PropTypes.node,
   navNext: PropTypes.node,
@@ -63,7 +68,8 @@ const propTypes = forbidExtraProps({
   onPrevMonthClick: PropTypes.func,
   onNextMonthClick: PropTypes.func,
   onOutsideClick: PropTypes.func,
-  renderDay: PropTypes.func,
+  renderCalendarDay: PropTypes.func,
+  renderDayContents: PropTypes.func,
   renderCalendarInfo: PropTypes.func,
 
   // accessibility
@@ -73,7 +79,9 @@ const propTypes = forbidExtraProps({
 
   // i18n
   monthFormat: PropTypes.string,
+  weekDayFormat: PropTypes.string,
   phrases: PropTypes.shape(getPhrasePropTypes(DayPickerPhrases)),
+  dayAriaLabelFormat: PropTypes.string,
 
   isRTL: PropTypes.bool,
 });
@@ -87,7 +95,6 @@ const defaultProps = {
   onClose() {},
 
   keepOpenOnDateSelect: false,
-  minimumNights: 1,
   isOutsideRange() {},
   isDayBlocked() {},
   isDayHighlighted() {},
@@ -102,7 +109,11 @@ const defaultProps = {
   withPortal: false,
   hideKeyboardShortcutsPanel: false,
   initialVisibleMonth: null,
+  firstDayOfWeek: null,
   daySize: DAY_SIZE,
+  verticalHeight: null,
+  noBorder: false,
+  transitionDuration: undefined,
 
   navPrev: null,
   navNext: null,
@@ -113,7 +124,8 @@ const defaultProps = {
   minDate: undefined,
   maxDate: undefined,
 
-  renderDay: null,
+  renderCalendarDay: undefined,
+  renderDayContents: null,
   renderCalendarInfo: null,
 
   // accessibility
@@ -123,6 +135,7 @@ const defaultProps = {
 
   // i18n
   monthFormat: 'MMMM YYYY',
+  weekDayFormat: 'dd',
   phrases: DayPickerPhrases,
 
   isRTL: false,
@@ -184,31 +197,46 @@ export default class DayPickerSingleDateController extends React.Component {
     } = nextProps;
     let { visibleDays } = this.state;
 
+    let recomputeOutsideRange = false;
+    let recomputeDayBlocked = false;
+    let recomputeDayHighlighted = false;
+
     if (isOutsideRange !== this.props.isOutsideRange) {
       this.modifiers['blocked-out-of-range'] = day => isOutsideRange(day);
+      recomputeOutsideRange = true;
     }
 
     if (isDayBlocked !== this.props.isDayBlocked) {
       this.modifiers['blocked-calendar'] = day => isDayBlocked(day);
+      recomputeDayBlocked = true;
     }
 
     if (isDayHighlighted !== this.props.isDayHighlighted) {
       this.modifiers['highlighted-calendar'] = day => isDayHighlighted(day);
+      recomputeDayHighlighted = true;
     }
-
+    
     // Mirai: New example to set custom classes for days 
     if (assignImportantCalendarClass !== this.props.assignImportantCalendarClass) {
       this.modifiers['important-calendar-class'] = day => assignImportantCalendarClass(day);
     }
 
+    const recomputePropModifiers = (
+      recomputeOutsideRange || recomputeDayBlocked || recomputeDayHighlighted
+    );
+
     if (
-      initialVisibleMonth !== this.props.initialVisibleMonth ||
       numberOfMonths !== this.props.numberOfMonths ||
-      enableOutsideDays !== this.props.enableOutsideDays
+      enableOutsideDays !== this.props.enableOutsideDays ||
+      (
+        initialVisibleMonth !== this.props.initialVisibleMonth &&
+        !this.props.focused &&
+        focused
+      )
     ) {
       const newMonthState = this.getStateForNewMonth(nextProps);
-      const currentMonth = newMonthState.currentMonth;
-      visibleDays = newMonthState.visibleDays;
+      const { currentMonth } = newMonthState;
+      ({ visibleDays } = newMonthState);
       this.setState({
         currentMonth,
         visibleDays,
@@ -225,7 +253,7 @@ export default class DayPickerSingleDateController extends React.Component {
       modifiers = this.addModifier(modifiers, date, 'selected');
     }
 
-    if (didFocusChange) {
+    if (didFocusChange || recomputePropModifiers) {
       values(visibleDays).forEach((days) => {
         Object.keys(days).forEach((day) => {
           const momentObj = moment(day);
@@ -235,22 +263,28 @@ export default class DayPickerSingleDateController extends React.Component {
             modifiers = this.deleteModifier(modifiers, momentObj, 'blocked');
           }
 
-          if (isOutsideRange(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'blocked-out-of-range');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-out-of-range');
+          if (didFocusChange || recomputeOutsideRange) {
+            if (isOutsideRange(momentObj)) {
+              modifiers = this.addModifier(modifiers, momentObj, 'blocked-out-of-range');
+            } else {
+              modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-out-of-range');
+            }
           }
 
-          if (isDayBlocked(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'blocked-calendar');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-calendar');
+          if (didFocusChange || recomputeDayBlocked) {
+            if (isDayBlocked(momentObj)) {
+              modifiers = this.addModifier(modifiers, momentObj, 'blocked-calendar');
+            } else {
+              modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-calendar');
+            }
           }
 
-          if (isDayHighlighted(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'highlighted-calendar');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'highlighted-calendar');
+          if (didFocusChange || recomputeDayHighlighted) {
+            if (isDayHighlighted(momentObj)) {
+              modifiers = this.addModifier(modifiers, momentObj, 'highlighted-calendar');
+            } else {
+              modifiers = this.deleteModifier(modifiers, momentObj, 'highlighted-calendar');
+            }
           }
           
           // Mirai: New example to set custom classes for days 
@@ -298,7 +332,7 @@ export default class DayPickerSingleDateController extends React.Component {
 
     onDateChange(day);
     if (!keepOpenOnDateSelect) {
-      onFocusChange({ focused: null });
+      onFocusChange({ focused: false });
       onClose({ date: day });
     }
   }
@@ -354,7 +388,7 @@ export default class DayPickerSingleDateController extends React.Component {
       },
     });
 
-    onPrevMonthClick();
+    onPrevMonthClick(prevMonth.clone());
   }
 
   onNextMonthClick() {
@@ -369,15 +403,16 @@ export default class DayPickerSingleDateController extends React.Component {
     const nextMonth = currentMonth.clone().add(numberOfMonths, 'month');
     const nextMonthVisibleDays = getVisibleDays(nextMonth, 1, enableOutsideDays);
 
+    const newCurrentMonth = currentMonth.clone().add(1, 'month');
     this.setState({
-      currentMonth: currentMonth.clone().add(1, 'month'),
+      currentMonth: newCurrentMonth,
       visibleDays: {
         ...newVisibleDays,
         ...this.getModifiers(nextMonthVisibleDays),
       },
     });
 
-    onNextMonthClick();
+    onNextMonthClick(newCurrentMonth.clone());
   }
 
 
@@ -399,7 +434,9 @@ export default class DayPickerSingleDateController extends React.Component {
       }
 
       const viableDays = days.filter(day => !this.isBlocked(day) && isAfterDay(day, focusedDate));
-      if (viableDays.length > 0) focusedDate = viableDays[0];
+      if (viableDays.length > 0) {
+        ([focusedDate] = viableDays);
+      }
     }
 
     return focusedDate;
@@ -444,11 +481,19 @@ export default class DayPickerSingleDateController extends React.Component {
     }
 
   getStateForNewMonth(nextProps) {
-    const { initialVisibleMonth, date, numberOfMonths, enableOutsideDays } = nextProps;
+    const {
+      initialVisibleMonth,
+      date,
+      numberOfMonths,
+      enableOutsideDays,
+    } = nextProps;
     const initialVisibleMonthThunk = initialVisibleMonth || (date ? () => date : () => this.today);
     const currentMonth = initialVisibleMonthThunk();
-    const visibleDays =
-      this.getModifiers(getVisibleDays(currentMonth, numberOfMonths, enableOutsideDays));
+    const visibleDays = this.getModifiers(getVisibleDays(
+      currentMonth,
+      numberOfMonths,
+      enableOutsideDays,
+    ));
     return { currentMonth, visibleDays };
   }
 
@@ -627,15 +672,26 @@ export default class DayPickerSingleDateController extends React.Component {
       enableOutsideDays,
       hideKeyboardShortcutsPanel,
       daySize,
-      renderDay,
+      firstDayOfWeek,
+      renderCalendarDay,
+      renderDayContents,
       renderCalendarInfo,
       isFocused,
       isRTL,
+      phrases,
+      dayAriaLabelFormat,
+      onOutsideClick,
+      onBlur,
+      showKeyboardShortcuts,
+      weekDayFormat,
+      verticalHeight,
+      noBorder,
+      transitionDuration,
     } = this.props;
 
-    const { phrases, currentMonth, visibleDays } = this.state;
+    const { currentMonth, visibleDays } = this.state;
 
-    return (
+    const dayPickerComponent = (
       <DayPicker
         orientation={orientation}
         enableOutsideDays={enableOutsideDays}
@@ -651,21 +707,41 @@ export default class DayPickerSingleDateController extends React.Component {
         hidden={!focused}
         hideKeyboardShortcutsPanel={hideKeyboardShortcutsPanel}
         initialVisibleMonth={() => currentMonth}
+        firstDayOfWeek={firstDayOfWeek}
         navPrev={navPrev}
         navNext={navNext}
         navPrevLocked={this.isMonthBlockedByMinDate()}
         navNextLocked={this.isMonthBlockedByMaxDate()}
         renderMonth={renderMonth}
-        renderDay={renderDay}
+        renderCalendarDay={renderCalendarDay}
+        renderDayContents={renderDayContents}
         renderCalendarInfo={renderCalendarInfo}
         isFocused={isFocused}
         getFirstFocusableDay={this.getFirstFocusableDay}
-        onBlur={this.onDayPickerBlur}
+        onBlur={onBlur}
         phrases={phrases}
         daySize={daySize}
         isRTL={isRTL}
+        showKeyboardShortcuts={showKeyboardShortcuts}
+        weekDayFormat={weekDayFormat}
+        dayAriaLabelFormat={dayAriaLabelFormat}
+        verticalHeight={verticalHeight}
+        noBorder={noBorder}
+        transitionDuration={transitionDuration}
       />
     );
+
+    if (onOutsideClick) {
+      return (
+        <OutsideClickHandler
+          onOutsideClick={onOutsideClick}
+        >
+          {dayPickerComponent}
+        </OutsideClickHandler>
+      );
+    }
+
+    return dayPickerComponent;
   }
 }
